refactor: extract code for static session launcher state
